requirements.txt
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
opencv-python==4.11.0.86
pillow==11.1.0
playwright==1.52.0